# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LifePath 2.0 API — The App That Fixes What Life Breaks
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: Get today's task list
      parameters:
        - in: query
          name: date
          schema:
            type: string
          description: ISO date string (defaults to today)
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task (complete, snooze, edit)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /mood:
    post:
      operationId: logMood
      tags: [mood]
      summary: Log morning mood check-in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoodInput"
      responses:
        "201":
          description: Mood logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoodLog"
  /mood/today:
    get:
      operationId: getTodayMood
      tags: [mood]
      summary: Get today's mood log if exists
      responses:
        "200":
          description: Today's mood
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoodLog"
        "404":
          description: No mood logged today

  # ─── WORK SESSIONS ────────────────────────────────────────────────────────
...
  /rooms:
    get:
      operationId: getRooms
      tags: [body-doubling]
      summary: Get available body doubling rooms
      responses:
        "200":
          description: Rooms list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
    post:
      operationId: createRoom
      tags: [body-doubling]
      summary: Create a new focus room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/{id}/join:
    post:
      operationId: joinRoom
      tags: [body-doubling]
      summary: Join a body doubling room
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinRoomInput"
      responses:
        "200":
          description: Joined room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomSession"
  /rooms/{id}/leave:
    post:
      operationId: leaveRoom
      tags: [body-doubling]
      summary: Leave a body doubling room
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Left room
...
components:
  schemas:
...
    Task:
      type: object
      required: [id, title, status, estimatedMinutes, isTwoMinute, date, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        microSteps:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, completed, missed, snoozed]
        estimatedMinutes:
          type: ["integer", "null"]
        isTwoMinute:
          type: boolean
        implementationIntention:
          type: ["string", "null"]
        temptationBundle:
          type: ["string", "null"]
        date:
          type: string
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    TaskInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        estimatedMinutes:
          type: integer
        implementationIntention:
          type: string
        temptationBundle:
          type: string
        scheduledDate:
          type: string
    TaskUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [completed, missed, snoozed, pending]
        title:
          type: string
        implementationIntention:
          type: string
        temptationBundle:
          type: string
...
    MoodLog:
      type: object
      required: [id, energy, focus, mood, date, createdAt]
      properties:
        id:
          type: integer
        energy:
          type: integer
          minimum: 1
          maximum: 5
        focus:
          type: integer
          minimum: 1
          maximum: 5
        mood:
          type: integer
          minimum: 1
          maximum: 5
        notes:
          type: ["string", "null"]
        date:
          type: string
        createdAt:
          type: string
    MoodInput:
      type: object
      required: [energy, focus, mood]
      properties:
        energy:
          type: integer
          minimum: 1
          maximum: 5
        focus:
          type: integer
          minimum: 1
          maximum: 5
        mood:
          type: integer
          minimum: 1
          maximum: 5
        notes:
          type: string

    # ── Focus Sessions ───────────────────────────────────────────────────────
...
    Room:
      type: object
      required: [id, name, presenceCount, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        presenceCount:
          type: integer
        isPrivate:
          type: boolean
        isActive:
          type: boolean
        currentSessionMinutes:
          type: ["integer", "null"]
        createdAt:
          type: string
    RoomInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        isPrivate:
          type: boolean
    JoinRoomInput:
      type: object
      required: [intention, durationMinutes]
      properties:
        intention:
          type: string
        durationMinutes:
          type: integer
    RoomSession:
      type: object
      required: [roomId, sessionId, joinedAt]
      properties:
        roomId:
          type: integer
        sessionId:
          type: integer
        joinedAt:
          type: string
...